7608 py2 fix (#5)
We only build PySide2 binaries for Python 3, the aim being Unreal. However, the framework ships some hooks which are used in other DCCs, where PySide is available natively, and run Python 2 (e.g. Maya 2020).
Allowed the framework to be loaded, even if a PySide2 build is not available for the current Python, if PySide or PySide2 is already available.